Move toward opening simple FLTK window.
Still having issues with getting this to compile/link correctly. It
sounds like the library /c/cygwin64/lib/w32api/libws2_32.a needs to
mentioned in the linker's command after the objects to be linked, but
this is not NetBeans' default order and I haven't figured out how to
change that without doing the makefiles entirely by hand, which seems a
huge pain.