Fix configuration to compile correctly
The previous configuration did not use the appropriate header files or
libraries to cross compile from cygwin to windows. This has now been
fixed.

Additionally, I learned it is extremely important to use the correct
*.dll files; no similar-but-slightly-different dll from a different
release of cygwin. This required culling some unnecessary entries in
PATH.

This still does not run from NetBeans correctly, but it does run from
the command line or just double-clicking the compiled *.exe.